20190506v2

+++ src/utils/file_utils.py
@@ -3,23 +3,26 @@ from shutil import copyfile, rmtree
 import shutil
 
 def read(file_path):
+    if not os.path.exists(file_path):
+        return
     with open(file_path, 'r') as f:
         # print(f.read())
         return(f.read())
 
 def write(file_path, content):
     pardir = os.path.dirname(file_path)
-    isExists = os.path.exists(pardir)
-    if not isExists:
+    if not os.path.exists(pardir):
         os.makedirs(pardir)
     with open(file_path, 'w') as f:
         f.write(content)
         
 def copy(src, dist):
-    copyfile(src, dist)
+    if os.path.exists(src):
+        copyfile(src, dist)
     
 def rename(src, dist):
-    os.rename(src, dist)
+    if os.path.exists(src):
+        os.rename(src, dist)
 
 def removeOld(path):
     if os.path.isdir(path):
@@ -28,7 +31,8 @@ def removeOld(path):
     os.remove(path)
     
 def remove(path):
-    rmtree(path)
+    if os.path.exists(path):
+        rmtree(path)
     
 def get_filename(file_path):
     return (os.path.split(file_path)[1])
